requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
mysql-connector-python==8.0.26
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
